server/requirements.txt
bcrypt
cryptography
email-validator
fastapi
motor
passlib
python-dotenv
python-jose
python-multipart